# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/product/terms/{id}": {
            "get": {
                "tags": [
                    "Product Terms"
                ],
                "summary": "Show Trade Item Offer",
                "operationId": "App\\Http\\Controllers\\TradeItemOfferController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Selected Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "basic": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Product Terms"
                ],
                "summary": "Update an existing Trade Item Offer",
                "operationId": "App\\Http\\Controllers\\TradeItemOfferController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Of Trade Item Offer To Update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "internal_name": {
                                        "description": "Internal Name",
                                        "type": "string"
                                    },
                                    "supplier_id": {
                                        "description": "Supplier Id",
                                        "type": "integer"
                                    },
                                    "supplier_trade_item_number": {
                                        "description": "Supplier Trade Item Number",
                                        "type": "string"
                                    },
                                    "customer_group_id": {
                                        "description": "Customer Group Id",
                                        "type": "integer"
                                    },
                                    "net_price": {
                                        "description": "Net Price",
                                        "type": "string"
                                    },
                                    "currency": {
                                        "description": "Currency",
                                        "type": "string"
                                    },
                                    "sales_unit": {
                                        "description": "Sales Unit",
                                        "type": "string"
                                    },
                                    "stock_keeping_unit": {
                                        "description": "Stock Keeping Unit",
                                        "type": "string"
                                    },
                                    "old_stock_keeping_unit": {
                                        "description": "Old Stock Keeping Unit",
                                        "type": "string"
                                    },
                                    "maximum_delivery_time": {
                                        "description": "Maximum Delivery Time",
                                        "type": "string"
                                    },
                                    "minimum_delivery_time": {
                                        "description": "Minimum Delivery Time",
                                        "type": "string"
                                    },
                                    "delivery_time_unit": {
                                        "description": "Delivery Time Unit",
                                        "type": "string"
                                    },
                                    "maximum_order_quantity": {
                                        "description": "Maximum Order Quantity",
                                        "type": "string"
                                    },
                                    "minimum_order_quantity": {
                                        "description": "Minimum Order Quantity",
                                        "type": "string"
                                    },
                                    "import_status": {
                                        "description": "Import status",
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "description": "Is Active",
                                        "type": "string"
                                    },
                                    "is_warehouse_item": {
                                        "description": "Is Warehouse Item",
                                        "type": "string"
                                    },
                                    "created_at ": {
                                        "description": "Created at",
                                        "type": "string",
                                        "format": "datetime"
                                    },
                                    "updated_at": {
                                        "description": "Updated at",
                                        "type": "string",
                                        "format": "datetime"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "basic": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product Terms"
                ],
                "summary": "Delete Trade Item Offer",
                "description": "Deletes trade item offer",
                "operationId": "deleteTradeItemoffer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Offer Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "basic": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "basic": {
                "type": "http",
                "scheme": "basic"
            }
        }
    }
}